% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageInAndOut.R
\name{buildInputCsv}
\alias{buildInputCsv}
\title{buildInputCsv}
\usage{
buildInputCsv(fastaPath, readPath, mode = "fastq", out = "~/MGCSIn")
}
\arguments{
\item{fastaPath}{A path or a vector of paths pointing to a directory containing only the reference sequences as fasta files with one of the following endings: .fasta, .fna, .fa.}

\item{readPath}{A path or a vector of paths pointing to a directory containing all read data to be used for setting up a RealReadSim file system. If it is a vector of paths, each paths reads are treated as the reads of different samples. For each read path there also has to be a fasta path. This is true even when all fasta paths are the same.}

\item{mode}{This parameter can be set to be "fastq", which will set the function to searching for fastq files and "bam", which will set the function to searching for bam files. The default is "fastq".}

\item{out}{The path to and the name of the input file without the suffix.}
}
\description{

}
