% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageInAndOut.R
\name{updateDataSystem}
\alias{updateDataSystem}
\title{updateDataSystem}
\usage{
updateDataSystem(oldPath, newPath)
}
\arguments{
\item{oldPath}{The path to the current directory of the datasystem.}

\item{newPath}{The path to the new location the datasystem shall be put to.}
}
\description{
This function moves a datasystem created by the MGCS::addToDataSystem function from oldPath to newPath and updates all paths in the datasystem.
}
